chain.getRuntimeVersion -> state.getRuntimeVersion
