+++ packages/rpc-core/src/jsonrpc.types.ts
@@ -22,10 +22,8 @@ export interface RpcInterface {
     getBlockHash(blockNumber?: BlockNumber | Uint8Array | number | string): Observable<Hash>;
     getFinalizedHead(): Observable<Hash>;
     getHeader(hash?: Hash | Uint8Array | string): Observable<Header>;
-    getRuntimeVersion(hash?: Hash | Uint8Array | string): Observable<RuntimeVersion>;
     subscribeFinalizedHeads(): Observable<Header>;
     subscribeNewHeads(): Observable<Header>;
-    subscribeRuntimeVersion(): Observable<RuntimeVersion>;
   };
   state: {
     call(method: Text | string, data: Bytes | Uint8Array | string, block?: Hash | Uint8Array | string): Observable<Bytes>;
@@ -40,6 +38,7 @@ export interface RpcInterface {
     getStorageHash(key: any, block?: Hash | Uint8Array | string): Observable<Hash>;
     getStorageSize(key: any, block?: Hash | Uint8Array | string): Observable<u64>;
     queryStorage(keys: (any)[], startBlock: Hash | Uint8Array | string, block?: Hash | Uint8Array | string): Observable<Vec<StorageChangeSet>>;
+    subscribeRuntimeVersion(): Observable<RuntimeVersion>;
     subscribeStorage<T = Codec[]>(keys: any[]): Observable<T>;
   };
   system: {
